Created new project "FTP" for the command line functionality and made minor changes to existing code for deleting directories remotely.
==DETAILS

For the CLI I opted to create a new project so that the GUI and CLI can be in
different installers (however, this can easily be changed). This is the
initial work for the CLI, the setup is complete.
The changes to the existing delete remote directories functions were made to
couple it with the existing delete remote file button.

==STORY

https://trello.com/c/FwM8Ig20